requirements.txt
google-genai
python-dotenv
requests==2.32.3
Flask==3.0.2
reportlab==4.1.0
pypdf==5.4.0